+++ src/main/java/org/snpsift/SnpSift.java
@@ -9,7 +9,6 @@ import org.snpeff.SnpEff;
 import org.snpeff.fileIterator.VcfFileIterator;
 import org.snpeff.snpEffect.Config;
 import org.snpeff.snpEffect.VcfAnnotator;
-import org.snpeff.snpEffect.commandLine.CommandLine;
 import org.snpeff.util.Download;
 import org.snpeff.util.Gpr;
 import org.snpeff.util.Log;
@@ -23,7 +22,7 @@ import org.snpsift.hwe.SnpSiftCmdHwe;
  *
  * @author pablocingolani
  */
-public class SnpSift implements VcfAnnotator, CommandLine {
+public class SnpSift implements VcfAnnotator {
 
 	// Version info (in sync with SnpEff)
 	public static final String BUILD = SnpEff.BUILD;
