+++ src/main/java/org/snpsift/annotate/mem/arrays/BoolArray.java
@@ -1,4 +1,4 @@
-package org.snpsift.annotate.mem;
+package org.snpsift.annotate.mem.arrays;
 
 import java.io.Serializable;
 import java.util.Arrays;
