+++ src/main/java/org/snpsift/annotate/mem/arrays/PosIndex.java
@@ -1,4 +1,4 @@
-package org.snpsift.annotate.mem;
+package org.snpsift.annotate.mem.arrays;
 
 import java.util.Arrays;
