+++ src/test/java/org/snpsift/tests/unit/TestCasesPosIndex.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
-import org.snpsift.annotate.mem.PosIndex;
+import org.snpsift.annotate.mem.arrays.PosIndex;
 
 public class TestCasesPosIndex {
